api/requirements.txt
aiohttp==3.7.4
aiodns==3.0.0
cchardet==2.1.7
python-decouple==3.4
